typescript refactor
