+++ ava.config.mjs
@@ -1,0 +1,6 @@
+export default {
+  extensions: {
+    ts: 'module',
+  },
+  nodeArguments: ['--loader=ts-node/esm'],
+};
