+++ src/parse.ts
@@ -1,0 +1,5 @@
+import { parse as parseAstro } from '@astrojs/parser';
+
+const parse = (text: string) => parseAstro(text);
+
+export default parse;
